[Feat] location 서비스 및 컨트롤러 작성, Menu 중복 방지

+++ src/main/java/com/sparta/project/controller/MenuController.java
@@ -10,6 +10,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
+import com.sparta.project.util.PermissionValidator;
 
 @RestController
 @RequiredArgsConstructor
@@ -17,6 +18,7 @@ import org.springframework.web.bind.annotation.*;
 public class MenuController {
 
     private final MenuService menuService;
+    private final PermissionValidator permissionValidator;
 
     // 음식점 메뉴 조회(ALL)
     @GetMapping
@@ -42,7 +44,8 @@ public class MenuController {
     public ApiResponse<MenuResponse> createMenu(
             @RequestBody MenuCreateRequest menuCreateRequest,
             Authentication authentication) {
-        MenuResponse newMenu = menuService.createMenu(menuCreateRequest, authentication);
+        permissionValidator.checkPermission(authentication, "OWNER", "MANAGER", "MASTER");
+        MenuResponse newMenu = menuService.createMenu(menuCreateRequest);
         return ApiResponse.success(newMenu);
     }
 
@@ -52,7 +55,8 @@ public class MenuController {
             @PathVariable String menu_id,
             @RequestBody MenuUpdateRequest menuUpdateRequest,
             Authentication authentication) {
-        MenuResponse updatedMenu = menuService.updateMenu(menu_id, menuUpdateRequest, authentication);
+        permissionValidator.checkPermission(authentication, "OWNER", "MANAGER", "MASTER");
+        MenuResponse updatedMenu = menuService.updateMenu(menu_id, menuUpdateRequest);
         return ApiResponse.success(updatedMenu);
     }
 
@@ -61,6 +65,7 @@ public class MenuController {
     public ApiResponse<Void> deleteMenu(
             @PathVariable String menu_id,
             Authentication authentication) {
+        permissionValidator.checkPermission(authentication, "OWNER", "MANAGER", "MASTER");
         menuService.deleteMenu(menu_id, authentication);
         return ApiResponse.success();
     }
